Use default options for db in resolvers.
The wrapper fn doesn't do anything special, just returns the same db each time.
It's only a fn so component metadata can be attached.

If we pass around the db itself instead, we can use with defaults to set the
options in one place instead of each call.

The server now manages default options as part of it's config, so the db is
passed through application context. This also lets the schema be a plain
function with no dependencies, so doesn't have to be a component and there isn't
that mess of bindings in closures.